Refactor C/U/D todo API handlers to separate file
This also exports the functions to be visible from other modules, should
they become necessary in the future.
