+++ src/api/todo.ts
@@ -1,7 +1,8 @@
-import { TodoApiItem } from "../types/todo";
+import { TodoApiItem, TodoItem } from "../types/todo";
 import { BACKEND_URL } from "../utils/constants";
 import { creatTodoFromApiParams } from "../utils/todo";
 
+// list + read; TODO: Decouple
 export async function fetchTasks(token: string | undefined) {
   if (!token) {
     return [];
@@ -18,4 +19,52 @@ export async function fetchTasks(token: string | undefined) {
   // TODO: Handle errors pending refactor
   const tasks = data.map(creatTodoFromApiParams);
   return tasks;
+}
+
+// create
+export async function addTodo(todo: TodoItem, token: string) {
+  const todoString = JSON.stringify({
+    ...todo,
+    tag: todo.tags?.join(", "),
+    reminderDate: todo.reminderDate?.toISOString(),
+  });
+  const requestOptions = {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${token}`,
+    },
+    body: todoString,
+  };
+  await fetch(`${BACKEND_URL}/todos`, requestOptions);
+}
+
+// update
+export async function updateTodo(todo: TodoItem, token: string) {
+  const todoString = JSON.stringify({
+    ...todo,
+    tag: todo.tags?.join(", "),
+    date: todo.reminderDate?.toISOString(),
+  });
+  const requestOptions = {
+    method: "PUT",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${token}`,
+    },
+    body: todoString,
+  };
+  await fetch(`${BACKEND_URL}/todos/${todo.id}`, requestOptions);
+}
+
+// delete
+export async function deleteTodo(todo: TodoItem, token: string) {
+  const requestOptions = {
+    method: "DELETE",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${token}`,
+    },
+  };
+  await fetch(`${BACKEND_URL}/todos/${todo.id}`, requestOptions);
 }
